steps: Move osrelease fixture to the fixtures module
Moves the fixture out of environment.py to its own module in
steps.fixtures.osrelease. The steps code shouldn't be importing from
environment.py.
